+++ examples/complete/outputs.tf
@@ -1,0 +1,3 @@
+output "alb_url" {
+  value = "https://${module.alb.load_balancer.lb_dns_name}"
+}