+++ examples/minimal/outputs.tf
@@ -1,8 +1,32 @@
-output "null_resource_id" {
-  value       = module.minimal.null_resource_id
-  description = "An arbitrary value that changes each time the resource is replaced."
+output "vpc_id" {
+  value = module.tamr_networking.vpc_id
 }
-output "example_value" {
-  value       = module.minimal.example_value
-  description = "Example variable."
+
+output "vpc_cidr_block" {
+  value = module.tamr_networking.vpc_cidr_block
+}
+
+
+output "compute_subnet_id" {
+  value = module.tamr_networking.compute_subnet_id
+}
+
+output "application_subnet_id" {
+  value = module.tamr_networking.application_subnet_id
+}
+
+output "application_subnet_cidr_block" {
+  value = module.tamr_networking.application_subnet_cidr_block
+}
+
+output "data_subnet_ids" {
+  value = module.tamr_networking.data_subnet_ids
+}
+
+output "tamr_ec2_availability_zone" {
+  value = module.tamr_networking.tamr_ec2_availability_zone
+}
+
+output "public_subnet_ids" {
+  value = module.tamr_networking.public_subnet_ids
 }