+++ examples/minimal/variables.tf
@@ -1,0 +1,4 @@
+variable "availability_zones" {
+  type        = list(string)
+  description = "The list of availability zones where we should deploy resources"
+}