+++ modules/alb/outputs.tf
@@ -1,0 +1,7 @@
+output "load_balancer" {
+  value = module.alb
+}
+
+output "lb_security_group_id" {
+  value = module.sg_https_lb.security_group_id
+}