+++ CHANGELOG.md
@@ -1,2 +1,4 @@
-# Tamr Terraform Template Repo - v0.1.0 - Feb 25th 2020
-* Initing project
+# Tamr Terraform AWS Networking Repo
+## v0.1.0 - July 20th 2021
+* Added minimal VPC, NACLs, S3 VPC Endpoint and optional public subnets and NAT gateway.
+* Added example with ALB and optional TLS Certificate